Begin removing non common functions
